type-c-service: Make `proxy` a directory module
Move existing proxy code into this directory.
